Add performance endpoints (#5589)
* scaffolding -- result type and db method shape

* add endpoints and endpoint scope test

* make sure perf requesting user can validate

using that as a stand-in for "manage" right now

* implement and test query for stats

* Update changelog

* More limit -> `Long` fallout

I didn't want to change methods in Dao, which would affect literally
every db subproject file, so there are a few ad hoc `toInt`s here.
Based on our usage patterns I don't expect to _need_ the `Long`-ness of
the different params, so uh this should be fine 😨

+++ app-backend/datamodel/src/main/scala/PaginatedResponse.scala
@@ -13,9 +13,11 @@ import io.circe.generic.JsonCodec
   * @param results sequence of results for a page
   */
 @JsonCodec
-final case class PaginatedResponse[A](count: Int,
-                                      hasPrevious: Boolean,
-                                      hasNext: Boolean,
-                                      page: Int,
-                                      pageSize: Int,
-                                      results: Seq[A])
+final case class PaginatedResponse[A](
+    count: Long,
+    hasPrevious: Boolean,
+    hasNext: Boolean,
+    page: Long,
+    pageSize: Long,
+    results: Seq[A]
+)
